+++ crates/rome_service/src/configuration/diagnostics.rs
@@ -289,6 +289,7 @@ mod test {
     use crate::{Configuration, MatchOptions, Matcher};
     use rome_deserialize::json::deserialize_from_json_str;
     use rome_diagnostics::{print_diagnostic_to_string, DiagnosticExt, Error};
+    use rome_json_parser::JsonParserOptions;
 
     fn snap_diagnostic(test_name: &str, diagnostic: Error) {
         let content = print_diagnostic_to_string(&diagnostic);
@@ -340,7 +341,8 @@ mod test {
     #[test]
     fn deserialization_error() {
         let content = "{ \n\n\"formatter\" }";
-        let result = deserialize_from_json_str::<Configuration>(content);
+        let result =
+            deserialize_from_json_str::<Configuration>(content, JsonParserOptions::default());
 
         assert!(result.has_errors());
         for diagnostic in result.into_diagnostics() {
@@ -363,6 +365,8 @@ mod test {
     }
   }
 }"#;
-        let _result = deserialize_from_json_str::<Configuration>(content).into_deserialized();
+        let _result =
+            deserialize_from_json_str::<Configuration>(content, JsonParserOptions::default())
+                .into_deserialized();
     }
 }